requirements.txt
python-telegram-bot>=21.3
APScheduler>=3.10.4
pytz>=2024.1
httpx>=0.27
python-docx>=1.1
